concertando bugs e reestruturando arquivos e objetos
