+++ Assets/scripts/completo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c3124309a9a9f5c42ab7a01e6311ddb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
