+++ Assets/scripts/completo/Destroy_obstacles.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+public class Destroy_obstacles : MonoBehaviour
+{
+    
+    // Just destroy watever touches this object
+     void OnCollisionEnter2D(Collision2D coll) 
+     {
+        Destroy(coll.gameObject);
+     }
+}
