+++ Assets/scripts/completo/Destroy_obstacles.cs.meta
