+++ Assets/scripts/completo/Game_Manager.cs
@@ -22,7 +22,7 @@ public class Game_Manager : MonoBehaviour{
 
     void Update()
     {
-        Debug.Log(phase);
+
         // Phase Query
         if(Time.time > phase_plan[phase] & phase < phase_plan.Length){
             phase++;
