+++ Assets/scripts/incompleto.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e6cda34a422e49428326a74a315564f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
