+++ Assets/scripts/incompleto/add_manager.cs
