+++ Assets/scripts/incompleto/dynamic_background_handler.cs
