+++ Assets/scripts/misc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e97e652ab4199d94b913bb973d619346
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
