--- Assets/scripts/obj.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3ed56519f5f774d479a7512dcfe8f418
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
