+++ Assets/sprites/Game_rules/end_handler.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3aca6894932e6644f94a1e1888da5b7b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
