+++ Assets/sprites/Game_rules/save_options.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1d604b25e16c21e43ad21ae82b56b8c2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
